Add comprehensive unit tests for splurge_test_namer
- Implemented tests for CLI functionalities including valid and invalid sentinel names, root imports, and command-line argument parsing.
- Created tests for exception handling in the namer module, ensuring proper error raising for target collisions, invalid inputs, and file operations.
- Developed tests for parser functions to validate sentinel reading and import finding, including handling of dynamic imports and annotated types.
- Added utility tests for file operations, ensuring robust error handling for reading, writing, and renaming files.
- Enhanced test coverage for edge cases, such as long sentinel names and fallback normalization.
