+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "splurge-test-namer"
-version = "2025.0.0"
+version = "2025.1.0"
 description = "Rename test modules based on sentinel metadata and optional import-following logic"
 readme = "README.md"
 license = "MIT"
@@ -14,6 +14,7 @@ classifiers = [
 	"Programming Language :: Python :: 3.10",
 	"Programming Language :: Python :: 3.11",
 	"Programming Language :: Python :: 3.12",
+	"Programming Language :: Python :: 3.13",
 ]
 
 [project.urls]
@@ -42,7 +43,11 @@ line-length = 120
 [tool.mypy]
 python_version = "3.10"
 ignore_missing_imports = true
-exclude = "^tests/"
+exclude = "^tests/|^build/|^dist/|^\\.venv/"
+files = ["splurge_test_namer"]
+
+# Ignore test modules (they intentionally contain test-only type patterns)
+
 
 [tool.setuptools.packages.find]
 where = ["."]
