+++ tests/unit/test_cli_0006.py
