+++ tests/unit/test_exceptions_namer_0002.py
