+++ tests/unit/test_exceptions_utils_0002.py
