+++ tests/unit/test_namer_0002.py
@@ -1,29 +1,15 @@
-import pytest
-
-from splurge_test_namer.namer import apply_renames
-from splurge_test_namer.exceptions import SplurgeTestNamerError
+from splurge_test_namer.namer import slug_sentinel_list
 
 
-def test_apply_renames_target_collision(tmp_path):
-    a = tmp_path / "tests" / "test_a.py"
-    b = tmp_path / "tests" / "test_b.py"
-    a.parent.mkdir(parents=True)
-    a.write_text("x")
-    b.write_text("y")
-
-    # propose renaming a -> existing file b (collision where b exists and is not in originals)
-    proposals = [(a, b)]
-    # Since b exists and is not an original in proposals, should raise
-    with pytest.raises(SplurgeTestNamerError):
-        apply_renames(proposals)
+def test_slug_replaces_spaces_and_dashes():
+    vals = ["Alpha Beta", "gamma-delta", "OK"]
+    slug = slug_sentinel_list(vals)
+    # spaces and dashes become underscores, lowercased
+    # case is preserved by the current policy
+    assert slug == "Alpha_Beta_gamma_delta_OK"
 
 
-def test_apply_renames_success(tmp_path):
-    a = tmp_path / "tests" / "test_a.py"
-    a.parent.mkdir(parents=True)
-    a.write_text("x")
-    target = a.with_name("test_misc_0001.py")
-    proposals = [(a, target)]
-    apply_renames(proposals)
-    assert target.exists()
-    assert not a.exists()
+def test_slug_trims_repeats_and_empty_parts():
+    vals = ["  ", "--a--", "b  c"]
+    slug = slug_sentinel_list(vals)
+    assert slug == "a_b_c"
