+++ tests/unit/test_parser_0005.py
@@ -3,63 +3,60 @@ from pathlib import Path
 from splurge_test_namer.parser import aggregate_sentinels_for_test
 
 
-def write_module(path: Path, content: str):
+def write_module(path: Path, sentinel_values: list[str] | None = None, extra: str = ""):
     path.parent.mkdir(parents=True, exist_ok=True)
+    content = "\n"
+    if sentinel_values is not None:
+        content += f"DOMAINS = {sentinel_values!r}\n"
+    content += extra
     path.write_text(content)
 
 
-def test_relative_import_too_deep(tmp_path):
-    repo = tmp_path / "repo"
-    pkg = repo / "pkg"
-    pkg.mkdir(parents=True)
-    write_module(pkg / "__init__.py", "DOMAINS = ['root']\n")
-
-    # create a test at top-level tests dir (not inside repo) that attempts a
-    # relative import from deeper than the module location (should be skipped)
-    tests = tmp_path / "tests"
-    tests.mkdir()
-    f = tests / "test_rel_deep.py"
-    write_module(f, "from ..pkg import something\n")
-
-    got = aggregate_sentinels_for_test(f, "pkg", repo, "DOMAINS")
-    # nothing should be found because relative import couldn't be resolved
-    assert got == []
-
-
-def test_mixed___all___exports(tmp_path):
+def test_import_resolution_various(tmp_path):
     repo = tmp_path / "repo"
     pkg = repo / "pkg"
     sub = pkg / "sub"
     sub.mkdir(parents=True)
-    # module that defines __all__ limiting what is exported
-    write_module(sub / "modx.py", "DOMAINS = ['x']\n__all__ = ['x']\n")
-    # another module not exported
-    write_module(sub / "mody.py", "DOMAINS = ['y']\n")
 
+    # Create modules with sentinel values
+    write_module(pkg / "__init__.py", ["root"])
+    write_module(sub / "mod.py", ["submod"])
+    write_module(sub / "star_mod.py", ["star"])
+
+    # test files that import these modules in different ways
     tests = tmp_path / "tests"
     tests.mkdir()
-    # star import from package: should still find modx but not mody if __all__ hides it
-    f = tests / "test_star_all.py"
-    write_module(f, "from pkg.sub import *\n")
 
-    got = aggregate_sentinels_for_test(f, "pkg", repo, "DOMAINS")
-    # Our parser doesn't evaluate __all__, but star-expansion will pick up files
-    assert "x" in got
-    assert "y" in got
+    # absolute import
+    f_abs = tests / "test_abs.py"
+    write_module(f_abs, None, "import pkg.sub.mod\n")
 
+    # from-import
+    f_from = tests / "test_from.py"
+    write_module(f_from, None, "from pkg.sub import mod\n")
 
-def test_namespace_package_resolution(tmp_path):
-    repo = tmp_path / "repo"
-    # Create namespace package (no __init__.py)
-    pkg = repo / "ns_pkg"
-    sub = pkg / "subpkg"
-    sub.mkdir(parents=True)
-    write_module(sub / "mod.py", "DOMAINS = ['ns']\n")
+    # relative import (simulate test inside package)
+    # create a test module inside pkg for relative imports
+    testpkg = pkg / "tests"
+    testpkg.mkdir()
+    f_rel = testpkg / "test_rel.py"
+    # from .sub import mod
+    write_module(f_rel, None, "from ..sub import mod\n")
 
-    tests = tmp_path / "tests"
-    tests.mkdir()
-    f = tests / "test_ns.py"
-    write_module(f, "from ns_pkg.subpkg import mod\n")
+    # star import
+    f_star = tests / "test_star.py"
+    write_module(f_star, None, "from pkg.sub import *\n")
 
-    got = aggregate_sentinels_for_test(f, "ns_pkg", repo, "DOMAINS")
-    assert "ns" in got
+    # Now aggregate from each test file and ensure values are found
+    got_abs = aggregate_sentinels_for_test(f_abs, "pkg", repo, "DOMAINS")
+    assert "submod" in got_abs
+
+    got_from = aggregate_sentinels_for_test(f_from, "pkg", repo, "DOMAINS")
+    assert "submod" in got_from
+
+    got_rel = aggregate_sentinels_for_test(f_rel, "pkg", repo, "DOMAINS")
+    assert "submod" in got_rel
+
+    got_star = aggregate_sentinels_for_test(f_star, "pkg", repo, "DOMAINS")
+    # star import should still pick up the module's sentinel via module resolution
+    assert "star" in got_star
