+++ tests/unit/test_parser_utils_0001.py
@@ -1,31 +1,32 @@
-from splurge_test_namer.parser import aggregate_sentinels_for_test
+import textwrap
+
+from splurge_test_namer.parser import read_sentinels_from_file
 from splurge_test_namer.util_helpers import resolve_module_to_paths
 
 
-def test_aggregate_handles_unreadable_module(tmp_path, monkeypatch):
-    repo = tmp_path / "repo"
-    tests = repo / "tests"
-    pkg = repo / "splurge_sql_tool"
-    pkg.mkdir(parents=True)
-    tests.mkdir(parents=True)
-
-    core = pkg / "core.py"
-    core.write_text("DOMAINS = ['core']\n")
-
-    t1 = tests / "test_badread.py"
-    t1.write_text("import splurge_sql_tool.core\n")
-
-    # make resolve_module_to_paths return the core path
-    paths = resolve_module_to_paths("splurge_sql_tool.core", repo)
-    assert any(p.samefile(core) for p in paths)
-
-    # monkeypatch the parser's safe_file_reader (it was imported there) to raise FileReadError
-    from splurge_test_namer.exceptions import FileReadError
-
-    monkeypatch.setattr(
-        "splurge_test_namer.parser.safe_file_reader", lambda p: (_ for _ in ()).throw(FileReadError("boom"))
+def test_read_sentinels_regex_fallback(tmp_path):
+    p = tmp_path / "tests" / "test_regex.py"
+    p.parent.mkdir(parents=True)
+    # write a file where AST parsing might not pick up (odd formatting)
+    p.write_text(
+        textwrap.dedent("""
+    DOMAINS = [
+        'one',
+        "two",
+    ]
+    """)
     )
+    vals = read_sentinels_from_file(p, "DOMAINS")
+    assert sorted(vals) == ["one", "two"]
 
-    sent = aggregate_sentinels_for_test(t1, "splurge_sql_tool", repo, "DOMAINS")
-    # should handle the read error and return empty
-    assert sent == []
+
+def test_resolve_module_fallback_search(tmp_path):
+    repo = tmp_path / "repo"
+    repo.mkdir()
+    # create file deep in repo that ends with name 'utils.py'
+    deep = repo / "some" / "path" / "utils.py"
+    deep.parent.mkdir(parents=True)
+    deep.write_text("# utils")
+
+    found = resolve_module_to_paths("splurge_sql_tool.utils", repo)
+    assert any(p.samefile(deep) for p in found)
