+++ tests/unit/test_utils_0001.py
@@ -1,77 +1,41 @@
 from pathlib import Path
-import pytest
 
-from splurge_test_namer.util_helpers import (
-    safe_file_reader,
-    safe_file_writer,
-    safe_file_renamer,
-    safe_file_rglob,
-    resolve_module_to_paths,
-)
-from splurge_test_namer.exceptions import FileReadError, FileWriteError, FileRenameError, FileGlobError
+from splurge_test_namer.util_helpers import safe_file_renamer
 
 
-def test_safe_file_reader_on_directory(tmp_path):
-    # reading a directory should raise FileReadError
-    with pytest.raises(FileReadError):
-        safe_file_reader(tmp_path)
-
-
-def test_safe_file_writer_parent_not_dir(tmp_path):
-    # create a file where the parent should be
-    parent = tmp_path / "parentfile"
-    parent.write_text("I'm a file")
-    dst = parent / "child.txt"
-    with pytest.raises(FileWriteError):
-        safe_file_writer(dst, "data")
-
-
-def test_safe_file_renamer_parent_not_dir(tmp_path):
-    parent = tmp_path / "parentfile"
-    parent.write_text("I am a file")
-    src = tmp_path / "a.txt"
-    src.write_text("a")
-    dst = parent / "b.txt"
-    with pytest.raises(FileRenameError):
-        safe_file_renamer(src, dst)
-
-
-def test_safe_file_renamer_overwrite_false(tmp_path):
+def test_safe_file_renamer_basic(tmp_path: Path) -> None:
     src = tmp_path / "a.txt"
     dst = tmp_path / "b.txt"
-    src.write_text("a")
-    dst.write_text("b")
-    with pytest.raises(FileRenameError):
-        safe_file_renamer(src, dst, overwrite=False)
+    src.write_text("hello")
+
+    safe_file_renamer(src, dst)
+
+    assert not src.exists()
+    assert dst.exists()
+    assert dst.read_text() == "hello"
 
 
-def test_safe_file_rglob_raises(monkeypatch, tmp_path):
-    # monkeypatch Path.rglob to raise
-    orig = Path.rglob
+def test_safe_file_renamer_prevent_overwrite(tmp_path: Path) -> None:
+    src = tmp_path / "a.txt"
+    dst = tmp_path / "b.txt"
+    src.write_text("one")
+    dst.write_text("two")
 
-    def fake_rglob(self, pattern):
-        raise RuntimeError("boom")
-
-    monkeypatch.setattr(Path, "rglob", fake_rglob)
+    # Should raise when destination exists and overwrite=False
     try:
-        with pytest.raises(FileGlobError):
-            safe_file_rglob(tmp_path, "*.py")
-    finally:
-        monkeypatch.setattr(Path, "rglob", orig)
+        safe_file_renamer(src, dst)
+        raised = False
+    except Exception:
+        raised = True
+    assert raised
 
 
-def test_resolve_module_to_paths_file_and_pkg(tmp_path):
-    repo = tmp_path / "repo"
-    pkg = repo / "pkg"
-    sub = pkg / "sub"
-    sub.mkdir(parents=True)
-    # create module file and package __init__
-    mod = sub / "mod.py"
-    mod.write_text("# mod")
-    init = sub / "__init__.py"
-    init.write_text("# init")
+def test_safe_file_renamer_overwrite_true(tmp_path: Path) -> None:
+    src = tmp_path / "a.txt"
+    dst = tmp_path / "b.txt"
+    src.write_text("one")
+    dst.write_text("two")
 
-    results = resolve_module_to_paths("pkg.sub.mod", repo)
-    assert any(p.name == "mod.py" for p in results)
-    results2 = resolve_module_to_paths("pkg.sub", repo)
-    assert any(p.name == "__init__.py" for p in results2)
+    # Allow overwrite
+    safe_file_renamer(src, dst, overwrite=True)
+    assert dst.read_text() == "one"
